+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Increased the limit of PitchTuning and PitchCorrection up from 1200 to 1800 cents https://github.com/oleg68/GrandOrgue/issues/33
 # 0.3.1.2338-8.os (2021-05-18)
 - Closing unnecessary jack clients https://github.com/oleg68/GrandOrgue/issues/30
 # 0.3.1.2338-6.os (2021-05-10)